[ISSUE #219]Supply detail information or read.me for eventmesh running in docker
