+++ docs/cn/instructions/eventmesh-runtime-quickstart.zh-CN.md
@@ -56,7 +56,7 @@ sh start.sh
 
 **2.3.1 项目结构说明：**
 
-![](../../images/project-structure.png)
+![project-structure](../../images/project-structure.png)
 - eventmesh-common : eventmesh公共类与方法模块
 - eventmesh-connector-api : eventmesh插件接口定义模块
 - eventmesh-connector-rocketmq : eventmesh rocketmq插件模块
@@ -157,15 +157,15 @@ docker run -d -p 10000:10000 -p 10105:10105 -v /data/eventmesh/rocketmq/conf/pro
 
 执行 `docker ps` 来检查容器的运行状况
 
-![image-20210309155917269](..\..\images\docker\docker-ps.png)
+![image-docker-ps](..\..\images\docker\docker-ps.png)
 
 执行 `docker logs [container id]` 可以得到如下结果
 
-![image-20210309195623836](..\..\images\docker\docker-logs.png)
+![image-docker-logs](..\..\images\docker\docker-logs.png)
 
 执行 `docker exec -it [container id] /bin/bash` 可以进入到容器中并查看详细信息
 
-![image-20210309200327627](..\..\images\docker\docker-exec.png)
+![image-docker-exec](..\..\images\docker\docker-exec.png)
 
 ### 3.4 测试
