+++ docs/cn/instructions/eventmesh-sdk-java-quickstart.zh-CN.md
@@ -10,7 +10,7 @@ TCP 和 Http 示例都在**eventmesh-test**模块下
 
 **注意：**下载了源代码后，需要将`/conf/application.properties` 和 `/conf/log4j2.xml` 复制到 `resources` 目录下
 
-![image-20210311193048090](..\..\images\eventmesh-test-structure.png)
+![image-test-structure](..\..\images\eventmesh-test-structure.png)
 
 ### 1. TCP DEMO
 
@@ -21,14 +21,14 @@ TCP 和 Http 示例都在**eventmesh-test**模块下
 - 启动消费者，订阅上一步骤已经创建的Topic
 
 ```
-运行com.webank.eventmesh.client.tcp.demo.AsyncSubscribe的主要方法
+运行com.webank.eventmesh.tcp.demo.AsyncSubscribe的主要方法
 ```
 
 - 启动发送端，发送消息
 
 
 ```
-运行com.webank.eventmesh.client.tcp.demo.AsyncPublish的主要方法
+运行com.webank.eventmesh.tcp.demo.AsyncPublish的主要方法
 ```
 
 <h4>广播消息</h4>
@@ -38,13 +38,13 @@ TCP 和 Http 示例都在**eventmesh-test**模块下
 - 启动消费端，订阅上一步骤已经创建的Topic
 
 ```
-运行com.webank.eventmesh.client.tcp.demo.AsyncSubscribeBroadcast的主要方法
+运行com.webank.eventmesh.tcp.demo.AsyncSubscribeBroadcast的主要方法
 ```
 
 - 启动发送端，发送广播消息
 
 ```
-运行com.webank.eventmesh.client.tcp.demo.AsyncPublishBroadcast的主要方法
+运行com.webank.eventmesh.tcp.demo.AsyncPublishBroadcast的主要方法
 ```
 
 ### 2. HTTP演示
@@ -62,13 +62,13 @@ TCP 和 Http 示例都在**eventmesh-test**模块下
   异步事件消费端为spring boot demo，运行demo即可启动服务并完成Topic订阅
 
 ```
-运行com.webank.eventmesh.client.http.demo.sub.SpringBootDemoApplication的主要方法
+运行com.webank.eventmesh.http.demo.sub.SpringBootDemoApplication的主要方法
 ```
 
 - 启动发送端，发送消息
 
 ```
-运行com.webank.eventmesh.client.http.demo.AsyncPublishInstance的主要方法
+运行com.webank.eventmesh.http.demo.AsyncPublishInstance的主要方法
 ```
