+++ eventmesh-test/src/main/java/com/webank/eventmesh/http/demo/SyncRequestInstance.java
@@ -1,21 +1,13 @@
-package com.webank.eventmesh.client.http.demo;
+package com.webank.eventmesh.http.demo;
 import com.webank.eventmesh.client.http.conf.LiteClientConfig;
 import com.webank.eventmesh.client.http.producer.LiteProducer;
 import com.webank.eventmesh.common.IPUtil;
 import com.webank.eventmesh.common.LiteMessage;
-import com.webank.eventmesh.common.ThreadPoolFactory;
 import com.webank.eventmesh.common.ThreadUtil;
-import com.webank.eventmesh.util.Utils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.LinkedList;
-import java.util.Properties;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
 
 public class SyncRequestInstance {
