+++ eventmesh-test/src/main/java/com/webank/eventmesh/http/demo/sub/SpringBootDemoApplication.java
@@ -1,4 +1,4 @@
-package com.webank.eventmesh.client.http.demo.sub;
+package com.webank.eventmesh.http.demo.sub;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
