+++ eventmesh-test/src/main/java/com/webank/eventmesh/tcp/demo/AsyncSubscribe.java
@@ -1,12 +1,12 @@
-package com.webank.eventmesh.client.tcp.demo;
+package com.webank.eventmesh.tcp.demo;
 
 import com.webank.eventmesh.client.tcp.WemqAccessClient;
-import com.webank.eventmesh.client.tcp.common.AccessTestUtils;
 import com.webank.eventmesh.client.tcp.common.ReceiveMsgHook;
 import com.webank.eventmesh.client.tcp.impl.DefaultWemqAccessClient;
 import com.webank.eventmesh.common.protocol.tcp.AccessMessage;
 import com.webank.eventmesh.common.protocol.tcp.Package;
 import com.webank.eventmesh.common.protocol.tcp.UserAgent;
+import com.webank.eventmesh.tcp.common.AccessTestUtils;
 import com.webank.eventmesh.util.Utils;
 import io.netty.channel.ChannelHandlerContext;
 import org.slf4j.Logger;
@@ -14,13 +14,13 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Properties;
 
-public class AsyncSubscribeBroadcast implements ReceiveMsgHook {
+public class AsyncSubscribe implements ReceiveMsgHook {
 
-    public static Logger logger = LoggerFactory.getLogger(AsyncSubscribeBroadcast.class);
+    public static Logger logger = LoggerFactory.getLogger(AsyncSubscribe.class);
 
     private static WemqAccessClient client;
 
-    public static AsyncSubscribeBroadcast handler = new AsyncSubscribeBroadcast();
+    public static AsyncSubscribe handler = new AsyncSubscribe();
 
     public static void main(String[] agrs)throws Exception{
         Properties properties = Utils.readPropertiesFile("application.properties");
@@ -32,7 +32,7 @@ public class AsyncSubscribeBroadcast implements ReceiveMsgHook {
             client.init();
             client.heartbeat();
 
-            client.subscribe("FT0-e-80030001-01-3");
+            client.subscribe("FT0-e-80010000-01-1");
             client.registerSubBusiHandler(handler);
 
             client.listen();
@@ -42,13 +42,13 @@ public class AsyncSubscribeBroadcast implements ReceiveMsgHook {
             //退出,销毁资源
 //            client.close();
         }catch (Exception e){
-            logger.warn("AsyncSubscribeBroadcast failed", e);
+            logger.warn("AsyncSubscribe failed", e);
         }
     }
 
     @Override
     public void handle(Package msg, ChannelHandlerContext ctx) {
         AccessMessage accessMessage = (AccessMessage)msg.getBody();
-        logger.info("receive broadcast msg==============={}", accessMessage);
+        logger.info("receive async msg====================={}", accessMessage);
     }
 }
